Fix poplar deprecation (#102)

+++ forest/bonsai/simulate_gps_data.py
@@ -1384,7 +1384,7 @@ def sim_gps_data(
         sys.stdout.write(f"	distance(km): {all_distances_array.tolist()}\n")
         sys.stdout.write(f"	hometime(hr): {all_times_array.tolist()}\n")
         obs = remove_data(all_traj, cycle, percentage, no_of_days)
-        obs_pd = prepare_data(obs, timestamp_s / 1000, tz_str)
+        obs_pd = prepare_data(obs, int(timestamp_s / 1000), tz_str)
         obs_pd['user'] = user + 1
         gps_data = pd.concat([gps_data, obs_pd])
